Collapse dependency recipes to have less at top level
Unlikely folks would run these separately, as it's only one dependency; if they would want to they can run `UpgradeDependencyVersion` separately.
