+++ src/main/resources/META-INF/rewrite/launchdarkly-6.yml
@@ -21,16 +21,8 @@ displayName: Migrate to LaunchDarkly 6.x
 description: This recipe will apply changes commonly needed when migrating to LaunchDarkly 6.x.
 recipeList:
   # https://docs.launchdarkly.com/sdk/server-side/java/migration-5-to-6
-  - org.openrewrite.launchdarkly.UpgradeLaunchDarkly6Dependencies
-  - org.openrewrite.launchdarkly.MigrateUserToContext
-
----
-type: specs.openrewrite.org/v1beta/recipe
-name: org.openrewrite.launchdarkly.UpgradeLaunchDarkly6Dependencies
-displayName: Migrate LaunchDarkly dependencies to 6.x
-description: Migrate LaunchDarkly dependencies to 6.x.
-recipeList:
   - org.openrewrite.java.dependencies.UpgradeDependencyVersion:
       groupId: com.launchdarkly
       artifactId: launchdarkly-java-server-sdk
       newVersion: 6.x
+  - org.openrewrite.launchdarkly.MigrateUserToContext
